ics3a/encoder: move irrv protocol header and pc file to top level
