+++ sources/encoder/irrv/CMakeLists.txt
@@ -1,0 +1,36 @@
+# Copyright (C) 2022 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions
+# and limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+add_library(irrv-protocol INTERFACE)
+target_include_directories(irrv-protocol INTERFACE ${CMAKE_SOURCE_DIR})
+
+if (IS_ABSOLUTE ${CMAKE_INSTALL_INCLUDEDIR})
+  set(PKGCONFIG_HEADER_DIR ${CMAKE_INSTALL_INCLUDEDIR})
+else()
+  set(PKGCONFIG_HEADER_DIR ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR})
+endif()
+
+if (IS_ABSOLUTE ${CMAKE_INSTALL_LIBDIR})
+  set(PKGCONFIG_LIB_DIR ${CMAKE_INSTALL_LIBDIR}/pkgconfig)
+else()
+  set(PKGCONFIG_LIB_DIR ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}/pkgconfig)
+endif()
+
+configure_file(${CMAKE_CURRENT_SOURCE_DIR}/irrv.pc.in ${CMAKE_CURRENT_BINARY_DIR}/irrv.pc @ONLY)
+
+install(FILES ${CMAKE_CURRENT_LIST_DIR}/irrv_protocol.h DESTINATION ${PKGCONFIG_HEADER_DIR}/irrv/)
+install(FILES ${CMAKE_CURRENT_BINARY_DIR}/irrv.pc DESTINATION ${PKGCONFIG_LIB_DIR})
+
