+++ sources/encoder/irrv/irrv.pc.in
